requirements.txt
ultralytics
paddleocr
opencv-python
streamlit
onnxruntime
numpy
PyYAML
